requirements.txt
azure-functions
azure-iot-device==2.0.0rc11
pyodbc